added global configuration, lots of fixes in RoutingM
other minor fixes

+++ Ex2/MessageTypes.h
@@ -4,7 +4,7 @@
 
 typedef struct BroadcastMsg
 {
-	uint8_t basestation_id;
+	uint16_t basestation_id;
 	uint16_t seq_nr;
 	uint16_t hop_count;
 	uint16_t sender_addr;
@@ -12,7 +12,7 @@ typedef struct BroadcastMsg
 
 typedef struct SimpleDataMsg
 {
-	uint8_t basestation_id;
+	uint16_t basestation_id;
 	uint16_t src_addr;
 	uint8_t data1;
 	uint8_t data2;
